Basic turns implemented
I think the previous commit was related to something I was trying with the packs for transformations. I don't really remember whether it worked or not at this point. That file is pretty gross, but I don't think it can be simplified significantly without trying to trick the borrow checker (which, I maintain, is my enemy)