Implemented smooth rotation
The remaining things I'd like to implement/tweak before taking a break is the standardization of the puzzle state between completed moves, and the replacement of SIMD macros with generic functions. I don't know what the former will look like for multi-rotation transformations, but those are still a ways out.